// FRONT/financial/src/types/generated/plannedTransactionDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0
 */
import type { CategoryDto } from './categoryDto';
import type { CurrencyDto } from './currencyDto';
import type { FrequencyDto } from './frequencyDto';
import type { SourceDtoLookup } from './sourceDtoLookup';
import type { TransactionTypeDto } from './transactionTypeDto';

export interface PlannedTransactionDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  amount?: number;
  startDate?: string;
  userId?: number;
  category?: CategoryDto;
  currency?: CurrencyDto;
  frequency?: FrequencyDto;
  source?: SourceDtoLookup;
  transactionType?: TransactionTypeDto;
}