// FRONT/financial/src/types/generated/createPlannedTransactionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0
 */

export interface CreatePlannedTransactionInput {
  /** @nullable */
  name?: string | null;
  amount?: number;
  startDate?: string;
  currencyId?: number;
  transactionTypeId?: number;
  /** @nullable */
  categoryId?: number | null;
  sourceId?: number;
  frequencyId?: number;
}