// FRONT/financial/src/types/generated/frequencyDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0
 */
import type { IntervalUnitDto } from './intervalUnitDto';

export interface FrequencyDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  userId?: number | null;
  intervalValue?: number;
  intervalUnit?: IntervalUnitDto;
}